Move app dependencies to requirements.txt
Many developers expect a requirements.txt file to be present in a
project in order to get a quick overview of the application's
dependencies and to be able to do a light-weight install of an
application for development via `pip install -r requirements.txt`.
Moving the dependencies out of setup.py also increases separation of
concerns in the project.